requirements.txt
ai21==3.0.1
Flask==2.0.1
google-generativeai==0.7.2
openai==0.27.0
python-dotenv
Pillow
requests
utils==1.0.2
Werkzeug==2.0.1
gunicorn